% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{common_words}
\alias{common_words}
\title{999 of the most common words in the English language}
\format{
A character vector with 999 elements, each representing a single word
commonly used in the English language.
}
\source{
\url{https://gist.github.com/deekayen/4148741}
}
\usage{
common_words
}
\description{
A data set containing 999 of the most common words in the English language.
Note that the source does not mention provenance of the list and has been
critiqued for diverging from typical lists of the most common words in the
English language; as such, this data set is referred to as "common words"
rather than "the most common words". The word "slave" has been omitted from
the list to avoid insensitive procedural name generation.
}
\keyword{datasets}
