% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{docker_adjectives}
\alias{docker_adjectives}
\title{Adjectives used in the procedural naming of Docker containers.}
\format{
A character vector with 108 elements, each representing a single
adjective used to name Docker containers.
}
\source{
\url{https://github.com/moby/moby/blob/master/pkg/namesgenerator/names-generator.go}
}
\usage{
docker_adjectives
}
\description{
A data set containing the 108 adjectives used in the naming of Docker
containers as of 2020-12-16.
}
\keyword{datasets}
