% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{docker_names}
\alias{docker_names}
\title{Surnames used in the procedural naming of Docker containers.}
\format{
A character vector with 237 elements, each representing a single
surname used to name Docker containers.
}
\source{
\url{https://github.com/moby/moby/blob/master/pkg/namesgenerator/names-generator.go}
}
\usage{
docker_names
}
\description{
A data set containing the 237 surnames used in the naming of Docker
containers as of 2020-12-16. Each name corresponds to a famous scientist,
technologist, or mathematician; see the original source for more information
on each name.
}
\keyword{datasets}
