% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proceduralnames-package.R
\docType{package}
\name{proceduralnames-package}
\alias{proceduralnames}
\alias{proceduralnames-package}
\title{proceduralnames: Several Methods for Procedural Name Generation}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A small, dependency-free way to generate random names. Methods 
    provided include the adjective-surname approach of Docker containers and 
    a list of common English words.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mikemahoney218/proceduralnames}
  \item Report bugs at \url{https://github.com/mikemahoney218/proceduralnames/issues}
}

}
\author{
\strong{Maintainer}: Michael Mahoney \email{mike.mahoney.218@gmail.com} (\href{https://orcid.org/0000-0003-2402-304X}{ORCID})

}
\keyword{internal}
