% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spanish_words}
\alias{spanish_words}
\title{820 of the most common words in the Spanish language}
\format{
A character vector with 820 elements, each representing a single word
commonly used in the Spanish language.
}
\source{
\url{http://corpus.rae.es/lfrecuencias.html}
}
\usage{
spanish_words
}
\description{
A data set containing 820 of the most common words in the Spanish language,
as determined by the RAE. Only words which can be represented in UTF-8 are
included for compatibility across platforms and collations.
}
\keyword{datasets}
